+++ CHANGES.md
@@ -1,3 +1,9 @@
+# v0.2 -- UNRELEASED
+
+Don't catch exceptions in `dispatch`.
+Split the `dispatch` function into `dispatch` and `dispatch*`.
+Added `locate-commands`, for when invoking `dispatch*` directly.
+
 # v0.1 -- 8 Aug 2022
 
 Initial release.